Feature: semi-random device_id generation. Storing&loading NVS flash
Some vibe coder code. Needs cleanup.

Added way generate a random device_id. ESP32 checks on boot if there is
one in NVS flash, if not it generates one and stores.

All topics are using this generated device_id.

Light State is also stored to NVS and checked on boot.